Correciones a ExtraService y paintwork. Los test de persistencia de Kind ya funcionan

+++ s2_galeriaarte-back/src/main/java/co/edu/uniandes/csw/galeriaarte/entities/ExtraServiceEntity.java
@@ -21,7 +21,7 @@ public class ExtraServiceEntity extends BaseEntity implements Serializable
 {
     private String name;
     private String description;
-    private double price;
+    private long price;
     private boolean availability;
     
     @PodamExclude
@@ -62,7 +62,7 @@ public class ExtraServiceEntity extends BaseEntity implements Serializable
     /**
      * @return the price
      */
-    public double getPrice()
+    public long getPrice()
     {
         return price;
     }
@@ -70,7 +70,7 @@ public class ExtraServiceEntity extends BaseEntity implements Serializable
     /**
      * @param price the price to set
      */
-    public void setPrice(double price)
+    public void setPrice(long price)
     {
         this.price = price;
     }
